Typechange

+++ src/param/param_slash.c
@@ -87,7 +87,7 @@ static void param_completer(struct slash *slash, char * token) {
 				prefix = param;
 				prefixlen = strlen(prefix->name);
 			} else {
-				int new_prefixlen = slash_prefix_length(prefix->name, param->name);
+				size_t new_prefixlen = slash_prefix_length(prefix->name, param->name);
 				if (new_prefixlen < prefixlen)
 					prefixlen = new_prefixlen;
 			}
